requirements/dev_requirements.txt
black==25.1.0
factory-boy==3.3.3
flake8==7.2.0
mock==5.2.0
pytest==8.3.5
pytest-flask-sqlalchemy==1.1.0
